+++ src/neighborhoods.jl
@@ -69,13 +69,32 @@ end
 
 function neighborhood(point::AbstractVector, tree,
                       ntype::FixedSizeNeighborhood, n::Int, w::Int = 1)
-    idxs = inrange(tree, point, ntype.ε)
+    idxs = NearestNeighbors.inrange(tree, point, ntype.ε)
     filter!((el) -> abs(el - n) ≥ w, idxs)
     return idxs
 end
 function neighborhood(point::AbstractVector, tree, ntype::FixedSizeNeighborhood)
-    idxs = inrange(tree, point, ntype.ε)
+    idxs = NearestNeighbors.inrange(tree, point, ntype.ε)
     return idxs
 end
 
 KDTree(D::AbstractDataset, metric::Metric = Euclidean()) = KDTree(D.data, metric)
+
+
+# TODO: This must use the new Neighborhood.jl
+"""
+    all_neighbors(vtree, vs, ns, K, w)
+Return the `maximum(K)`-th nearest neighbors for all input points `vs`, with indices `ns` in
+original data, while respecting the theiler window `w`.
+"""
+function all_neighbors(vtree, vs, ns, K, w)
+    k, sortres, N = maximum(K), true, length(vs)
+    dists = [Vector{eltype(vs[1])}(undef, k) for _ in 1:N]
+    idxs = [Vector{Int}(undef, k) for _ in 1:N]
+    for i in 1:N
+        # The skip predicate also skips the point itself for w ≥ 0
+        skip = j -> ns[i] - w ≤ j ≤ ns[i] + w
+        NearestNeighbors.knn_point!(vtree, vs[i], sortres, dists[i], idxs[i], skip)
+    end
+    return idxs, dists
+end