+++ test/runtests.jl
@@ -1,5 +1,15 @@
 using DelayEmbeddings
 
+# Download some test timeseries
+repo = "https://raw.githubusercontent.com/JuliaDynamics/ExercisesRepo/master/timeseries"
+tsfolder = joinpath(@__DIR__, "timeseries")
+todownload = ["$n.csv" for n in 1:4]
+
+mkpath(tsfolder)
+for a in todownload
+    download(repo*"/"*a, joinpath(tsfolder, a))
+end
+
 ti = time()
 
 const diffeq = (atol = 1e-9, rtol = 1e-9, maxiters = typemax(Int))
@@ -10,6 +20,7 @@ include("delaytime_test.jl")
 include("delaycount_test.jl")
 # include("test_pecora.jl")
 include("uzal_cost_test.jl")
+include("mdop_tests.jl")
 
 ti = time() - ti
 println("\nTest took total time of:")