+++ test/uzal_cost_test.jl
@@ -1,9 +1,8 @@
 using DynamicalSystemsBase
 using DelayEmbeddings
-using StatsBase
 using Test
-using Statistics
 using Random
+using DelimitedFiles
 
 println("\nTesting uzal_cost.jl...")
 @testset "Uzal cost" begin
@@ -45,44 +44,34 @@ println("\nTesting uzal_cost.jl...")
     @test L<-3
 end
 
-@testset "Lorenz system" begin
-    ## Simple Check on Lorenz System
-    lo = Systems.lorenz()
-
-    tr = trajectory(lo, 10; dt = 0.01, Ttr = 10)
-
-    # check Euclidean metric
-    L= uzal_cost(tr;
-        Tw = 60, K= 3, w = 12, samplesize = 1.0,
-        metric = Euclidean()
-    )
-    L_max = -2.411
-    L_min = -2.412
-    @test L_min < L < L_max
+@testset "Uzal local cost (Lorenz)" begin
+    tr = readdlm(joinpath(tsfolder, "3.csv"))
+    tr = Dataset(tr)
 
     # check local cost function output
     Tw = 60
-    L_local= uzal_cost_local(tr;
-        Tw = Tw, K= 3, w = 12,metric = Euclidean()
+    L = uzal_cost(tr;
+        Tw = Tw, K= 3, w = 12, samplesize = 1.0,
+        metric = Euclidean()
     )
+    L_local= uzal_cost_local(tr; Tw = Tw, K= 3, samplesize = 1.0,w = 12, metric = Euclidean())
     @test length(L_local) == length(tr)-Tw
     @test maximum(L_local)>L
     @test minimum(L_local)<L
-
-    # check Maximum metric
-    L = uzal_cost(tr;
-        Tw = Tw, K= 3, w = 12, samplesize = 1.0,
-        metric = Chebyshev()
-    )
-    L_max = -2.475
-    L_min = -2.485
 end
 
 @testset "Roessler system" begin
+    # For comparison reasons using Travis CI we carry out the integration on a UNIX
+    # OS and save the resulting time series
+
     ## Test Roessler example as in Fig. 7 in the paper with internal data
-    ro = Systems.roessler([1.0, 1.0, 1.0], a=0.15, b = 0.2, c=10)
-    tr = trajectory(ro, 1250; dt = 0.125, Ttr = 10)
-    x = tr[:,1]
+    # ro = Systems.roessler([1.0, 1.0, 1.0], a=0.15, b = 0.2, c=10)
+    # tr = trajectory(ro, 1250; dt = 0.2, Ttr = 100)
+    # writedlm("4.csv", tr)
+
+    tr = readdlm(joinpath(tsfolder, "4.csv"))
+    tr = Dataset(tr)
+    x = tr[:, 1]
 
     # theiler window
     w  = 12
@@ -111,47 +100,47 @@ end
         end
     end
 
-    tau_min = 9
-    tau_max = 11
+    tau_min = 4
+    tau_max = 8
 
     min1_idx = sortperm(L[1,:])
     min1 = tw_max[min1_idx[1]]
     @test tau_min < min1 < tau_max
-    L_min = -2.96
-    L_max = -2.8
+    L_min = -2.7
+    L_max = -2.3
     @test L_min < L[1,min1_idx[1]] < L_max
 
 
     min2_idx = sortperm(L[2,:])
     min2 = tw_max[min2_idx[1]]
     @test tau_min < min2 < tau_max
-    L_min = -2.64
-    L_max = -2.5
+    L_min = -2.3
+    L_max = -2.1
     @test L_min < L[2,min2_idx[1]] < L_max
 
 
     min3_idx = sortperm(L[3,:])
     min3 = tw_max[min3_idx[1]]
     @test tau_min < min3 < tau_max
-    L_min = -2.46
-    L_max = -2.3
+    L_min = -2.1
+    L_max = -2.0
     @test L_min < L[3,min3_idx[1]] < L_max
 
 
     min4_idx = sortperm(L[4,:])
     min4 = tw_max[min4_idx[1]]
     @test tau_min < min4 < tau_max
-    L_min = -2.36
-    L_max = -2.2
+    L_min = -2.0
+    L_max = -1.9
     @test L_min < L[4,min4_idx[1]] < L_max
 
     L_local = uzal_cost_local(tr;
-        Tw = Tw, K= 3, w = 12,
+        Tw = Tw, K= 3, w = 12, samplesize=1.0,
         metric = Chebyshev()
     )
     L = uzal_cost(tr;
         Tw = Tw, K= 3, w = 12,
-        metric = Chebyshev(),samplesize=1.0
+        metric = Chebyshev(), samplesize=1.0
     )
     @test length(L_local) == length(tr)-Tw
     @test maximum(L_local) > L
@@ -232,8 +221,9 @@ end
 
 ## Test Lorenz example as in Fig. 7 in the paper with internal data
 @testset "Lorenz fig. 7." begin
-    lo = Systems.lorenz([1.0, 1.0, 50.0])
-    tr = trajectory(lo, 100; dt = 0.01, Ttr = 10)
+    # lo = Systems.lorenz([0, 10.0, 0.0]; σ=10, ρ=28, β=8/3)
+    # tr = trajectory(lo, 100; dt = 0.01, Ttr = 100)
+    tr = readdlm(joinpath(tsfolder, "4.csv"))
     x = tr[:, 1]
 
     # theiler window
@@ -249,29 +239,27 @@ end
     # maximum neighbours
     k_max = 4
     # number of total trials
-    trials = 12
+    trials = 16
 
     # preallocation
     L = zeros(k_max,trials)
     tw_max = zeros(trials)
 
-    for K = 1:k_max
-        for i = 1:trials
-            tw_max[i] = i*(m-1)
-            Y = embed(x,m,i)
+    for i = 1:trials
+        tw_max[i] = i*(m-1)
+        Y = embed(x,m,i)
+        for K = 1:k_max
             L[K,i] = uzal_cost(Y; Tw=Tw, K=K, w=w, samplesize=SampleSize, metric=metric)
         end
     end
 
-    tau_min = 17
-    tau_max = 21
-
+    tau_min = 5
+    tau_max = 7
     min1_idx = sortperm(L[1,:])
     min1 = tw_max[min1_idx[1]]
     @test tau_min < min1 < tau_max
     L_max = -2.3
     @test L[1,min1_idx[1]] < L_max
-
 
     min2_idx = sortperm(L[2,:])
     min2 = tw_max[min2_idx[1]]
@@ -279,18 +267,16 @@ end
     L_max = -2.0
     @test L[2,min2_idx[1]] < L_max
 
-
     min3_idx = sortperm(L[3,:])
     min3 = tw_max[min3_idx[1]]
-    @test tau_min < min3 < tau_max
-    L_max = -1.95
+    @test 5 < min3 < 7
+    L_max = -1.9
     @test L[3,min3_idx[1]] < L_max
-
 
     min4_idx = sortperm(L[4,:])
     min4 = tw_max[min4_idx[1]]
-    @test tau_min < min4 < tau_max
-    L_max = -1.9
+    @test 5 < min4 < 7
+    L_max = -1.8
     @test L[4,min4_idx[1]] < L_max
 
     # # plot results using PyPlot
@@ -305,7 +291,7 @@ end
     # xlabel(L"$t_w$")
     # ylabel(L"$L_k$")
     # legend(labels)
-    # xticks(0:2:24)
+    # xticks(0:2:32)
     # #yscale("symlog")
     # title("Lorenz System as in Fig. 7(a) in the Uzal Paper")
     # grid()